Deprecate missing_math.h
Once we were fully C++11 and moved MSVS minimum to >= 2015, the contents
of this file were no longer needed as far as I can tell. What few things
were here that anybody might still rely on, are in fmath.h.

+++ src/iv/imageviewer.cpp
@@ -2076,7 +2076,7 @@ static inline void
 calc_subimage_from_zoom(const IvImage* img, int& subimage, float& zoom,
                         float& xcenter, float& ycenter)
 {
-    int rel_subimage = Imath::trunc(log2f(1 / zoom));
+    int rel_subimage = Imath::trunc(std::log2(1.0f / zoom));
     subimage         = clamp<int>(img->subimage() + rel_subimage, 0,
                           img->nsubimages() - 1);
     if (!(img->subimage() == 0 && zoom > 1)
